+++ sql/dbupdate.php
@@ -584,3 +584,25 @@ if ($ilDB->tableExists($table_name)) {
     ]);
 }
 ?>
+<#23>
+<?php
+/** @var $ilDB ilDBInterface */
+$table_name = 'srlcm_tokens';
+$column_name = 'event';
+
+if ($ilDB->tableColumnExists($table_name, $column_name)) {
+    $legacy_event_name_mapping = [
+        'onPostpone' => 'routine_postpone',
+        'onOptOut' => 'routine_opt_out',
+        'onDelete' => 'routine_delete',
+    ];
+
+    foreach ($legacy_event_name_mapping as $legacy_event => $new_event) {
+        $ilDB->update(
+            $table_name,
+            ['event' => ['text', $new_event]],
+            ['event' => ['text', $legacy_event]]
+        );
+    }
+}
+?>